experiments/text/multiclass/dair-ai-emotion/requirements.txt
numpy
datasets
transformers
torch
scikit-learn
tqdm
matplotlib
umap-learn
